Nuevas fuentes, seleccion de fuentes y fluidez entre ventanas

+++ unlpimage/src/windows/memes.py
@@ -1,4 +1,3 @@
-#Segunda natalla de memes
 import PySimpleGUI as sg
 import os
 import json
@@ -6,16 +5,12 @@ from src.default.pathing import BASE_PATH, templates_memes_path
 from src.default.data import read_config
 from src.default.data import read_memes 
 from src.default.data import get_img_data_profiles as get_img_data
-#from generar_etiquetas import arbol
-#from src.windows import generar_etiquetas
 from PIL import Image, ImageTk, ImageFont, ImageDraw, ImageFont
 from src.default.setup import *
 from src.windows.funciones_memes import tam_box, entra, calcular_tam_fuente, boxes_2, boxes_3
 from src import function_registo
 
-#archivo_memes = open('template_meme.json', 'r')
-#datos = json.load(archivo_memes)
-             
+
 ruta_repositorio, ruta_collages, ruta_memes = read_config()
 datos = read_memes()
 templates_path = templates_memes_path
@@ -23,7 +18,7 @@ templates_path = templates_memes_path
 
 ##############################################################################################
 
-def layout_memes_prueba():
+def layout_memes_templates():
     ruta_repositorio, ruta_collages, ruta_memes = read_config()
     col_1 =[
             [sg.Image(data=get_img_data(os.path.join(BASE_PATH,'src', 'default','memes-templates', 'Batman_golpea_robin.png'), first = True), key='Batman_golpea_robin.png', enable_events=True, metadata=0 ),
@@ -40,10 +35,13 @@ def layout_memes_prueba():
     layout = [[sg.Text("Seleccione un template para crear un meme.", **text_format15)],[sg.Column(col_1)], [sg.Column(col_2)],[sg.Button('Volver', key='-VOLVER-')]]
 
     return layout
+
 ############################################################################################### 
+
 def layout_crear_2box():
     col_1 = [
-            [sg.Text("Elegir tipografía"), sg.FileBrowse("Elegir tipografía",key='-FUENTE-', initial_folder=ruta_repositorio, file_types=(("TTF Files", "*.ttf"),))],
+            #[sg.Text("Elegir tipografía"), sg.FileBrowse("Elegir tipografía",key='-FUENTE-', initial_folder=ruta_repositorio, file_types=(("TTF Files", "*.ttf"),))],
+            [sg.Button("Elegir tipografía", key='-FUENTE-')],
             [sg.Text("Cuadro de texto 1")],
             [sg.InputText(key='-TEXTO1-')],
             [sg.Text("Cuadro de texto 2")],
@@ -60,9 +58,10 @@ def layout_crear_2box():
     return layout
 
 ############################################################################################### 
+
 def layout_crear_3box():
     col_1 = [
-            [sg.Text("Elegir tipografía"), sg.FileBrowse("Elegir tipografía",key='-FUENTE-', initial_folder=ruta_repositorio, file_types=(("TTF Files", "*.ttf"),))],
+            [sg.Button("Elegir tipografía", key='-FUENTE-')],
             [sg.Text("Cuadro de texto 1")],
             [sg.InputText(key='-TEXTO1-')],
             [sg.Text("Cuadro de texto 2")],
@@ -79,6 +78,24 @@ def layout_crear_3box():
 
     layout = [[sg.Column(col_1), sg.Column(col_2)]]
     return layout
+
+def layout_fuentes():
+    col_1 =[
+            [sg.Image(data=get_img_data(os.path.join(BASE_PATH,'src', 'default','flavor_sans.png'), first = True), key='-flavor-', enable_events=True, metadata=0 ),
+             sg.Image(data=get_img_data(os.path.join(BASE_PATH,'src','default','hawaianas.png'), first = True), key='-hawaianas-', enable_events=True, metadata=0 ) ]
+            ]
+    
+    
+#images column
+    col_2 = [
+            [sg.Image(data=get_img_data(os.path.join(BASE_PATH,'src','default','invisible.png'), first = True), key='-invisible-', enable_events=True, metadata=0 ),
+             sg.Image(data=get_img_data(os.path.join(BASE_PATH,'src','default','tommy.png'), first = True), key='-tommy-', enable_events=True, metadata=0 ) ]
+            ]
+    
+    layout = [[sg.Text("Seleccione un template para crear un meme.", **text_format15)],[sg.Column(col_1)], [sg.Column(col_2)],[sg.Button('Volver', key='-VOLVER-')]]
+
+    return layout
+
 ###############################################################################################
 
 def layout_mostrar():
@@ -95,19 +112,15 @@ def ventana_meme(perfil):
     Una vez presionado Generar, se guardará el meme nuevo.
    
     """
-    window = sg.Window('Generador de memes',layout_memes_prueba())
+    confirm = ''
+    window = sg.Window('Generador de memes',layout_memes_templates())
     while True:
         event1, values1 = window.read()
         if event1 == sg.WINDOW_CLOSED or event1 == '-VOLVER-':
             window.close()
             break
         if event1 == 'Batman_golpea_robin.png' or event1 == 'hide_the_pain_harold.png' or event1 == 'novio_mira_otra_mujer.png' or event1 == 'seguro_esta_pensando_en_otra.png':
-            #folder_path = 'BASE_PATH'
-            #file_types = [('Imagen', '*.png'), ('Imagen', '*.jpg')]
-            #file_path = sg.popup_get_file('Seleccionar imagen', initial_folder=ruta_repositorio, file_types=file_types)
-            #if file_path:
-            #    image_name = os.path.basename(file_path)
-            #    print("Imagen seleccionada:", image_name)
+
             file_path = os.path.join(BASE_PATH,'src', 'default','memes-templates', event1)
             image_name = event1
             color_relleno = (0, 0, 0)  # (R, G, B)
@@ -135,21 +148,53 @@ def ventana_meme(perfil):
             meme_final = meme_original.copy()
             draw = ImageDraw.Draw(meme_final)
             
+
             while True:
+                if confirm == 'No':
+                    window.close()
+                    break
                 window.Hide()
                 event2, values2 = window_boxes.read()
                 if event2 == sg.WINDOW_CLOSED or event2 == '-VOLVER-' or event2 == '-TEXTONO-':
                     window_boxes.close()
                     window.UnHide()
-                    break    
+                    break  
+                if event2 == '-FUENTE-':
+                    window_fuente = sg.Window('Seleccionar fuente', layout_fuentes())
+                    while True:
+                        event_f,values_f = window_fuente.read()
+                        match event_f:
+                            case('-flavor-'):
+                                fuente_elegida = os.path.join(BASE_PATH,'src', 'default','memes-templates','fuentes','Flavor_sans.ttf')
+                                window_fuente.close()
+                                break
+                            case('-hawaianas-'):
+                                fuente_elegida = os.path.join(BASE_PATH,'src', 'default','memes-templates','fuentes','Hawainas.ttf')
+                                window_fuente.close()
+                                break
+                            case('-invisible-'):
+                                fuente_elegida = os.path.join(BASE_PATH,'src', 'default','memes-templates','fuentes','Invisible.ttf')
+                                window_fuente.close()
+                                break
+                            case('-tommy-'):
+                                fuente_elegida = os.path.join(BASE_PATH,'src', 'default','memes-templates','fuentes','Tommy.ttf')
+                                window_fuente.close()
+                                break
+                        if event_f == sg.WINDOW_CLOSED or event_f == '-VOLVER-':
+                                window_fuente.close()
+                                break
             
-                elif event2 == '-TEXTOSI_2box-':
-                    fuente_elegida = values2['-FUENTE-']
+                if event2 == '-TEXTOSI_2box-':
+                   # fuente_elegida = values2['-FUENTE-']
                     texto1 = values2['-TEXTO1-']
                     texto2 = values2['-TEXTO2-']
                     print(texto1, texto2)
                     
-                    fuente = ImageFont.truetype(fuente_elegida, 200) #el tamaño
+                    try:
+                        fuente = ImageFont.truetype(fuente_elegida, 200) #el tamaño
+                    except UnboundLocalError:
+                        fuente_elegida = os.path.join(BASE_PATH,'src', 'default','memes-templates','fuentes','Flavor_sans.ttf')
+                        fuente = ImageFont.truetype(fuente_elegida, 200)
 
                     x1, y1, x2, y2 = resultado_funcion1[0]
                     x3, y3, x4, y4 = resultado_funcion1[1]
@@ -166,15 +211,18 @@ def ventana_meme(perfil):
                     draw.text((x1,y1), texto1, font=fuente_ajustada_1,fill=color_relleno)
                     draw.text((x3,y3), texto2, font=fuente_ajustada_2,fill=color_relleno)
                 
-                elif event2 == '-TEXTOSI_3box-':
-                    fuente_elegida = values2['-FUENTE-']
+                if event2 == '-TEXTOSI_3box-':
+                    #fuente_elegida = values2['-FUENTE-']
                     texto1 = values2['-TEXTO1-']
                     texto2 = values2['-TEXTO2-']
                     texto3 = values2['-TEXTO3-']
                     print(texto1, texto2, texto3)
-            
 
-                    fuente = ImageFont.truetype(fuente_elegida, 200) #el tamaño    
+                    try:
+                        fuente = ImageFont.truetype(fuente_elegida, 200) #el tamaño
+                    except UnboundLocalError:
+                        fuente_elegida = os.path.join(BASE_PATH,'src', 'default','memes-templates','fuentes','Flavor_sans.ttf')
+                        fuente = ImageFont.truetype(fuente_elegida, 200)  
                
                     x1, y1, x2, y2 = resultado_funcion1[0]
                     x3, y3, x4, y4 = resultado_funcion1[1] 
@@ -200,59 +248,65 @@ def ventana_meme(perfil):
                 Image.open(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','meme_final.png'))
             
                 window_mostrar = sg.Window('Meme final', layout_mostrar())
-                #window_boxes.hide()
-
-                while True:
-                    window_boxes.Hide()
-                    event3, values3 = window_mostrar.read()
-                    if event3 == sg.WINDOW_CLOSED or event3 == '-VOLVER2-':
-                        window_mostrar.close()
-                        window_boxes.UnHide()
-                        break
-                        #window_boxes.Unhide()
-                    elif event3 == '-GUARDARMEME-':
-                        confirm = sg.popup_yes_no('¿Desea guardar la imagen?', title='Guardar imagen')
-                        if confirm == 'No':
+                if event2 == '-TEXTOSI_2box-' or event2 == '-TEXTOSI_3box-':
+                    while True:
+                        window_boxes.Hide()
+                        event3, values3 = window_mostrar.read()
+                        if event3 == sg.WINDOW_CLOSED or event3 == '-VOLVER2-':
                             window_mostrar.close()
                             window_boxes.UnHide()
+                            os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','meme_final.png'))
                             texto1=None
                             texto2=None
                             texto3=None
+                            meme_final = meme_original.copy()
+                            draw = ImageDraw.Draw(meme_final)
                             break
-                        if confirm == 'Yes':
-                            filename = sg.popup_get_text('Ingrese el nombre de archivo:', title='Guardar imagen')
-
-                            if filename:
-                         # Aquí puedes escribir la lógica para guardar la imagen con el nombre de archivo ingresado
-                                print('Imagen guardada con el nombre:', filename)
-                                meme_final.save(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba', filename + '.png'))
-
-                                Action = 'Generó un meme'
-                                Values = image_name
-                                
-                                if image_name == "novio_mira_otra_mujer.png":
-                                    Texts = texto1, texto2, texto3
-                                else:
-                                    Texts = texto1, texto2
-                                function_registo(perfil, Action, Values, Texts)
+                        elif event3 == '-GUARDARMEME-':
+                            confirm = sg.popup_yes_no('¿Desea guardar la imagen?', title='Guardar imagen')
+                            if confirm == 'No':
                                 window_mostrar.close()
-                                window_boxes.close()
-                                window.UnHide()
-                                os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','meme_final.png'))
-                                os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','imagen_con_cuadros_de_texto.png'))
+                                window_boxes.UnHide()
+                                texto1=None
+                                texto2=None
+                                texto3=None
+                                meme_final = meme_original.copy()
+                                draw = ImageDraw.Draw(meme_final)
                                 break
-
-
-                   #window_boxes.UnHide()
-                
-            
-            
-            
-        #window.close()
+                            if confirm == 'Yes':
+                                filename = sg.popup_get_text('Ingrese el nombre de archivo:', title='Guardar imagen')
     
-    #window.close()
+                                if filename:
+                             # Aquí puedes escribir la lógica para guardar la imagen con el nombre de archivo ingresado
+                                    print('Imagen guardada con el nombre:', filename)
+                                    meme_final.save(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba', filename + '.png'))
     
+                                    Action = 'Generó un meme'
+                                    Values = [image_name]
+                                    
+                                    if image_name == "novio_mira_otra_mujer.png":
+                                        Texts = texto1+' '+texto2+' '+texto3
+                                    else:
+                                        Texts = texto1+' '+texto2
+                                    function_registo(perfil, Action, Values, Texts)
+                                    confirm = sg.popup_yes_no('¿Desea seguir generando memes?', title='Crear memes')
+                                    if confirm == 'Yes':
+                                        window_mostrar.close()
+                                        window_boxes.close()
+                                        window.UnHide()
+                                        os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','meme_final.png'))
+                                        os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','imagen_con_cuadros_de_texto.png'))
+                                        
+                                    if confirm == 'No':
+                                        window_mostrar.close()
+                                        window_boxes.close()
+                                        #window.close()
+                                        os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','meme_final.png'))
+                                        os.remove(os.path.join(BASE_PATH, 'src', 'Repositorio_prueba','imagen_con_cuadros_de_texto.png'))
+                                        
 
+                                    break
+    
 
 #
 #############################################################################
